+++ TodoApp.WebApp/Client/src/modules/category/category.component.ts
@@ -1,4 +1,4 @@
-import { Component, EventEmitter, OnInit, Output } from "@angular/core";
+import { Component, EventEmitter, OnInit, Output, OnDestroy } from "@angular/core";
 import { ICategoryList } from "./model/category.list.model";
 import { ICategoryItem } from "./model/categoryItem.model";
 import { CategoryService } from "./category.service";
@@ -10,6 +10,8 @@ import { IMessage } from "../shared/models/IMessage";
 import { Result } from "../shared/enums/Result";
 import { NotificationType } from "../shared/enums/NotificationType";
 import { NotificationPopupComponent } from "../shared/components/notification/notification.component";
+import { NgxSpinner, NgxSpinnerService } from "ngx-spinner";
+import { ConfirmationDialogComponent } from "../shared/components/confirmation-dialog/confirmation.component";
 
 @Component({
     selector: "todo-category",
@@ -18,9 +20,10 @@ import { NotificationPopupComponent } from "../shared/components/notification/no
 })
 
 
-export class CategoryComponent implements OnInit {
+export class CategoryComponent implements OnInit, OnDestroy {
 
     categories: ICategoryItem[] = [];
+    selected: number = 0;
 
     @Output()
     clickViewItem: EventEmitter<number> = new EventEmitter();
@@ -30,8 +33,9 @@ export class CategoryComponent implements OnInit {
 
     constructor(private categoryService: CategoryService,
         public dialog: MatDialog,
-        private _snackBar: MatSnackBar
-        ) {
+        private _snackBar: MatSnackBar,
+        private _spinner: NgxSpinnerService
+    ) {
     }
 
     ngOnInit(): void {
@@ -39,14 +43,16 @@ export class CategoryComponent implements OnInit {
     }
 
     async loadData(): Promise<void> {
+        this._spinner.show();
         var allItem = await this.categoryService.getAll();
         allItem.subscribe(item => {
             this.categories = item;
+            this._spinner.hide();
         });
     }
 
     viewActionItem(item: ICategoryItem): void {
-
+        this.selected = item.id;
         this.clickViewItem.emit(item.id);
     }
 
@@ -54,7 +60,8 @@ export class CategoryComponent implements OnInit {
         const dialogRef = this.dialog.open(AddTodoCategoryComponent, { height: '240px' });
 
         dialogRef.afterClosed().subscribe(result => {
-            console.log(result);
+            if (result.type)
+                this.loadData();
         });
     }
 
@@ -63,18 +70,33 @@ export class CategoryComponent implements OnInit {
     }
 
     async deleteCategory(item: ICategoryItem): Promise<void> {
-        var deleteRes = await this.categoryService.delete(item);
-        deleteRes.subscribe(res =>{
-            let notification: IMessage = {
-                type: res.result == Result.Error ? NotificationType.Error : NotificationType.Information,
-                message: res.result == Result.Success ? ["Remove succesfully"] : res.messages
-            };
-            this._snackBar.openFromComponent(NotificationPopupComponent, {
-                data: notification,
-                duration: 3000
+        const dialogRef = this.dialog.open(ConfirmationDialogComponent, { data: `Are you sure to delete category ${item.name}. This action could not be undo!` });
+
+        dialogRef.afterClosed().subscribe(async (result) => {
+            if (!result.type) return;
+
+            this._spinner.show();
+            var deleteRes = await this.categoryService.delete(item);
+            deleteRes.subscribe(res => {
+                let notification: IMessage = {
+                    type: res.result == Result.Error ? NotificationType.Error : NotificationType.Information,
+                    message: res.result == Result.Success ? ["Remove succesfully"] : res.messages
+                };
+                this._snackBar.openFromComponent(NotificationPopupComponent, {
+                    data: notification,
+                    duration: 3000
+                });
+                this.selected = 0;
+                this.loadData();
+                this.clickViewItem.emit(0);
             });
-            this.loadData();
-            this.clickViewItem.emit(0);
         });
+
+    }
+
+    ngOnDestroy(): void {
+        this.categories = [];
+        this.selected = 0;
+
     }
 }
